# mod1/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: A sample API built using OpenAPI and swagger-node-codegen. CRUD operations below. Use for guidance only and see the docs for more.
  version: 0.1.9
servers:
  - url: http://localhost:3000
    description: Local server
paths:
  /users:
    post:
      summary: Creates a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: string
                  description: The username that was added
        "400":
          description: Bad request. No request body provided.
    get:
      summary: Returns a list of users.
      description: Standard get request retriving all users
      responses:
        "200":
          description: A JSON array of user names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    patch:
      summary: Updates a user given a username and new username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                new_username:
                  type: string
          responses:
            "200":
              description: Updated
            "400":
              description: Bad request. No request body provided.
    delete:
      summary: Deletes a user given a username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
          responses:
            "200":
              description: Deleted